Refactor comiler
Declar, Statements, and expressions are now static functions referenced elsewhere
